LE CAST EN SHORT

+++ src/vm/instr/write_read_reg.c
@@ -6,7 +6,7 @@
 /*   By: sjimenez <sjimenez@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/03/14 21:52:27 by sjimenez          #+#    #+#             */
-/*   Updated: 2018/03/17 18:12:21 by hbouillo         ###   ########.fr       */
+/*   Updated: 2018/03/18 19:40:30 by bacrozat         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,7 +50,7 @@ t_value				get_parameter_result(t_par par, t_proc *process, int restricted)
 		result = read_reg(process->reg[par.value - 1]);
 	else if (par.type == T_ID)
 	{
-		read = read_memory(process->pc + process->owner->spawn, par.value, REG_SIZE, restricted);
+		read = read_memory(process->pc + process->owner->spawn, (short)par.value, REG_SIZE, restricted);
 		i = -1;
 		while (++i < REG_SIZE)
 		{
